Split shared vs internal tables for histology module.
